General Model for Codeigniter
Its a generalised library for model in codeingiter, where no need to make separate model for insert, update, delete and get single results and get multiple results again and again. Its a generalised library where you can use this throughout the project. 

Very flexible and easy to use everything mentioned in the comment how to use and why we used the parameters